+++ settings.php
@@ -240,6 +240,16 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
                     get_string('buttoncolorlightmoodleoutline', 'theme_boost_union', null, true),
         ];
 
+        // Prepare button size options.
+        $buttonsizeoptions = [
+            THEME_BOOST_UNION_SETTING_BUTTONSIZE_SMALL =>
+                    get_string('buttonsizesmall', 'theme_boost_union', null, true),
+            THEME_BOOST_UNION_SETTING_BUTTONSIZE_MEDIUM =>
+                    get_string('buttonsizemedium', 'theme_boost_union', null, true),
+            THEME_BOOST_UNION_SETTING_BUTTONSIZE_LARGE =>
+                    get_string('buttonsizelarge', 'theme_boost_union', null, true),
+        ];
+
         // Prepare login method divider type options.
         $logindivideroptions = [
             THEME_BOOST_UNION_SETTING_LOGINDIVIDERTYPE_NONE =>
@@ -1681,6 +1691,25 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
             THEME_BOOST_UNION_SETTING_SELECT_YES
         );
 
+        // Setting: Local login button size.
+        $name = 'theme_boost_union/loginlocalbuttonsize';
+        $title = get_string('loginlocalbuttonsizesetting', 'theme_boost_union', null, true);
+        $description = get_string('loginlocalbuttonsizesetting_desc', 'theme_boost_union', null, true);
+        $setting = new admin_setting_configselect(
+            $name,
+            $title,
+            $description,
+            THEME_BOOST_UNION_SETTING_BUTTONSIZE_MEDIUM,
+            $buttonsizeoptions
+        );
+        $tab->add($setting);
+        $page->hide_if(
+            'theme_boost_union/loginlocalbuttonsize',
+            'theme_boost_union/loginlocalloginenable',
+            'neq',
+            THEME_BOOST_UNION_SETTING_SELECT_YES
+        );
+
         // Setting: Local login divider type.
         $name = 'theme_boost_union/loginlocaldividertype';
         $title = get_string('loginlocaldividertypesetting', 'theme_boost_union', null, true);
@@ -1868,6 +1897,25 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $tab->add($setting);
         $page->hide_if(
             'theme_boost_union/loginidpbuttoncolor',
+            'theme_boost_union/loginidploginenable',
+            'neq',
+            THEME_BOOST_UNION_SETTING_SELECT_YES
+        );
+
+        // Setting: IDP login button size.
+        $name = 'theme_boost_union/loginidpbuttonsize';
+        $title = get_string('loginidpbuttonsizesetting', 'theme_boost_union', null, true);
+        $description = get_string('loginidpbuttonsizesetting_desc', 'theme_boost_union', null, true);
+        $setting = new admin_setting_configselect(
+            $name,
+            $title,
+            $description,
+            THEME_BOOST_UNION_SETTING_BUTTONSIZE_MEDIUM,
+            $buttonsizeoptions
+        );
+        $tab->add($setting);
+        $page->hide_if(
+            'theme_boost_union/loginidpbuttonsize',
             'theme_boost_union/loginidploginenable',
             'neq',
             THEME_BOOST_UNION_SETTING_SELECT_YES
@@ -2141,6 +2189,25 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
             THEME_BOOST_UNION_SETTING_SELECT_YES
         );
 
+        // Setting: Self registration button size.
+        $name = 'theme_boost_union/loginselfregistrationbuttonsize';
+        $title = get_string('loginselfregistrationbuttonsizesetting', 'theme_boost_union', null, true);
+        $description = get_string('loginselfregistrationbuttonsizesetting_desc', 'theme_boost_union', null, true);
+        $setting = new admin_setting_configselect(
+            $name,
+            $title,
+            $description,
+            THEME_BOOST_UNION_SETTING_BUTTONSIZE_MEDIUM,
+            $buttonsizeoptions
+        );
+        $tab->add($setting);
+        $page->hide_if(
+            'theme_boost_union/loginselfregistrationbuttonsize',
+            'theme_boost_union/loginselfregistrationenable',
+            'neq',
+            THEME_BOOST_UNION_SETTING_SELECT_YES
+        );
+
         // Setting: Self registration divider type.
         $name = 'theme_boost_union/loginfirsttimesignupdividertype';
         $title = get_string('loginfirsttimesignupdividertypesetting', 'theme_boost_union', null, true);
@@ -2322,6 +2389,25 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $tab->add($setting);
         $page->hide_if(
             'theme_boost_union/loginguestbuttoncolor',
+            'theme_boost_union/loginguestloginenable',
+            'neq',
+            THEME_BOOST_UNION_SETTING_SELECT_YES
+        );
+
+        // Setting: Guest login button size.
+        $name = 'theme_boost_union/loginguestbuttonsize';
+        $title = get_string('loginguestbuttonsizesetting', 'theme_boost_union', null, true);
+        $description = get_string('loginguestbuttonsizesetting_desc', 'theme_boost_union', null, true);
+        $setting = new admin_setting_configselect(
+            $name,
+            $title,
+            $description,
+            THEME_BOOST_UNION_SETTING_BUTTONSIZE_MEDIUM,
+            $buttonsizeoptions
+        );
+        $tab->add($setting);
+        $page->hide_if(
+            'theme_boost_union/loginguestbuttonsize',
             'theme_boost_union/loginguestloginenable',
             'neq',
             THEME_BOOST_UNION_SETTING_SELECT_YES